backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
google-generativeai
openai
pinecone-client
chromadb
sentence-transformers
PyMuPDF
python-docx